Refactor Client code
